% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{ARNT.metadata}
\alias{ARNT.metadata}
\title{Metadata data frame}
\format{
a data frame of one row and 7 variables.
}
\usage{
data("ARNT.metadata")
}
\description{
Used to run examples. Data frame containing metadata information
for the ChIP-Seq GSM2390643.
Fields in the data frame:
\itemize{
  \item Name: Name of the file.
  \item Accession: Accession ID of the experiment.
  \item Cell: Cell line or tissue.
  \item 'Cell Type': More information about the cells.
  \item Treatment
  \item Antibody
  \item TF: Transcription factor tested in the ChIP-Seq experiment.
}
}
\keyword{datasets}
