% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{ARNT.peaks.bed}
\alias{ARNT.peaks.bed}
\title{ChIP-Seq dataset}
\format{
a data frame of 2140 rows and 4 variables.
}
\usage{
data("ARNT.peaks.bed")
}
\description{
Used to run examples. Data frame containing peak information
from the ChIP-Seq GSM2390643.
Fields in the data frame:
\itemize{
  \item Name: Name of the file.
  \item chr: Chromosome, factor
  \item start: Start coordinate for each peak
  \item end: End coordinate for each peak
  \item X.10.log.pvalue.: log10(p-Value) for each peak.
}
}
\keyword{datasets}
