% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{chip_metadata}
\alias{chip_metadata}
\title{ChIP-seq transcription factor metadata}
\format{
A data frame with 1267 rows. Columns include:
\describe{
  \item{chip.name}{The TF name in the ChIP-seq experiment.}
  \item{tf.name}{The HGNC gene symbol of the TF.}
  \item{EntrezID}{Entrez IDs.}
}
}
\usage{
data("chip_metadata")
}
\description{
Metadata linking ChIP-seq experiments to their corresponding 
transcription factors. Each row provides the annotation of one TF.
}
\keyword{datasets}
