% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{ChIPDB}
\alias{ChIPDB}
\title{TF-gene binding binary matrix}
\format{
a matrix of 1060 columns and 16797 rows
}
\usage{
data("ChIPDB")
}
\description{
Its rows correspond to human genes, and its columns, to every ChIP-Seq
experiment in the database. The values are 1 – if the ChIP-Seq
has a peak assigned to that gene – or 0 – if it hasn’t –.
}
\keyword{datasets}
