% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{CM_list}
\alias{CM_list}
\title{List of contingency matrix}
\format{a list of 10 contingency matrix.}
\usage{
data("CM_list")
}
\description{
Used to run examples. List of 10 contingency matrix, output of
the function "contingency_matrix" from the TFEA.ChIP package.
}
\keyword{datasets}
