% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{contingency_matrix}
\alias{contingency_matrix}
\title{Computes 2x2 contingency matrices}
\usage{
contingency_matrix(
  test_list,
  control_list = NULL,
  chip_index = get_chip_index()
)
}
\arguments{
\item{test_list}{List of gene Entrez IDs for the test group.}

\item{control_list}{List of gene Entrez IDs for the control group. If not 
provided, all human genes not in test_list will be used as control.}

\item{chip_index}{Data frame containing accession IDs of ChIPs and their
associated transcription factors. Default is obtained from get_chip_index().}
}
\value{
A list of contingency matrices, one per element in chip_index.
}
\description{
Function to compute a 2x2 contingency matrix based on the
presence or absence of gene IDs in a ChIP-Seq binding database.
}
\examples{
data('Genes.Upreg', package = 'TFEA.ChIP')
CM_list_UP <- contingency_matrix(Genes.Upreg)
}
