% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{deseq.result}
\alias{deseq.result}
\title{deseq.result}
\format{DESeqResult object with 10 elements}
\usage{
data("deseq.result")
}
\description{
Used to run examples. part of a DESeqResults object.
}
\keyword{datasets}
