% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{deseq2table}
\alias{deseq2table}
\title{Extracts data from a DESeq2 results object.}
\usage{
deseq2table(deseq.result)
}
\arguments{
\item{deseq.result}{DESeqResults object. Must include gene IDs}
}
\value{
A table containing Entrez Gene IDs, LogFoldChange and p-val
values, sorted by log2FoldChange
}
\description{
Function to extract Gene IDs, logFoldChange, and p-val
values from a DESeqResults object. Gene IDs are translated to ENTREZ IDs, if possible.
}
\examples{
data("deseq.result",package="TFEA.ChIP")
deseq2table(deseq.result)
}
