% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{DnaseHS_db}
\alias{DnaseHS_db}
\title{DHS databse}
\format{GenomicRanges object with 32 elements}
\usage{
data("DnaseHS_db")
}
\description{
Used to run examples. Part of a DHS database storing 32 sites
for the human genome in GenomicRanges format.
}
\keyword{datasets}
