% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{Entrez.gene.IDs}
\alias{Entrez.gene.IDs}
\title{List of Entrez Gene IDs}
\format{Array of 2754 Entrez Gene IDs.}
\usage{
data("Entrez.gene.IDs")
}
\description{
Used to run examples. Array of 2754 Entrez Gene IDs extracted
from an RNA-Seq experiment sorted by log(Fold Change).
}
\keyword{datasets}
