% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{Genes.Upreg}
\alias{Genes.Upreg}
\title{List of Entrez Gene IDs}
\format{Array of 2754 Entrez Gene IDs.}
\usage{
data("Genes.Upreg")
}
\description{
Used to run examples. Array of 342 Entrez Gene IDs extracted from upregulated genes in an RNA-Seq experiment.
}
\keyword{datasets}
