% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{getCMstats}
\alias{getCMstats}
\title{Generate statistical parameters from a contingency_matrix output}
\usage{
getCMstats(contMatrix_list, chip_index = get_chip_index())
}
\arguments{
\item{contMatrix_list}{Output of “contingency_matrix”, a list of
contingency matrix.}

\item{chip_index}{Output of the function “get_chip_index”, a data frame
containing accession IDs of ChIPs on the database and the TF each one
tests. If not provided, the whole internal database will be used}
}
\value{
Data frame containing accession ID of a ChIP-Seq experiment, the
TF tested in that experiment, raw p-value (-10*log10 pvalue), odds-ratio
and FDR-adjusted p-values (-10*log10 adj.pvalue).
}
\description{
From a list of contingency matrices, such as the output
from “contingency_matrix”, this function computes a fisher's exact test
for each matrix and generates a data frame that stores accession ID of a
ChIP-Seq experiment, the TF tested in that experiment, the p-value and
the odds ratio resulting from the test.
}
\examples{
data("CM_list",package = "TFEA.ChIP")
stats_mat_UP <- getCMstats(CM_list)
}
