% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{gr.list}
\alias{gr.list}
\title{List of one ChIP-Seq dataset}
\format{List of one ChIP-Seq dataset.}
\usage{
data("gr.list")
}
\description{
Used to run examples. List of one ChIP-Seq dataset (from GSM2390643)
in GenomicRanges format with 2140 peaks.
}
\keyword{datasets}
