% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{GSEA_run}
\alias{GSEA_run}
\title{Function to run a GSEA analysis}
\usage{
GSEA_run(
  gene.list,
  LFC,
  chip_index = get_chip_index(),
  get.RES = FALSE,
  RES.filter = NULL,
  perms = 1000
)
}
\arguments{
\item{gene.list}{List of Entrez IDs ordered by their fold change.}

\item{LFC}{Vector of log2( Fold Change ) values.}

\item{chip_index}{Output of the function “get_chip_index”, a data frame
containing accession IDs of ChIPs on the database and the TF each one
tests. If not provided, the whole internal database will be used}

\item{get.RES}{(Optional) boolean. If TRUE, the function stores Running
Enrichment Scores of all/some TF.}

\item{RES.filter}{(Optional) chr vector. When get.RES==TRUE, allows to
choose which TF's Running Enrichment Score to store.}

\item{perms}{Number of permutations for a permutation test.}
}
\value{
a list of:
Enrichment.table: data frame containing accession ID, Cell type, ChIP-Seq
treatment, transcription factor tested, enrichment score, raw and
adjusted p-value, and argument of every ChIP-Seq experiment.
RES (optional): list of running sums of every ChIP-Seq
indicators (optional): list of 0/1 vectors that stores the matches (1)
and mismatches (0) between the gene list and the gene set.
}
\description{
Function to run a GSEA to analyze the distribution of TFBS
across a sorted list of genes.
}
\examples{
data('hypoxia',package = 'TFEA.ChIP')
hypoxia <- preprocessInputData(hypoxia)
chip_index<-get_chip_index(TFfilter = c('HIF1A','EPAS1','ARNT'))
GSEA.result <- GSEA_run( hypoxia$Genes, hypoxia$log2FoldChange, chip_index, get.RES = TRUE)
}
