% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{GSEA_Shuffling}
\alias{GSEA_Shuffling}
\title{Function to create shuffled gene lists to run GSEA.}
\usage{
GSEA_Shuffling(gene.list, permutations)
}
\arguments{
\item{gene.list}{Vector of gene Entrez IDs.}

\item{permutations}{Number of shuffled gene vector IDs required.}
}
\value{
Vector of randomly arranged Entrez IDs.
}
\description{
Function to create shuffled gene lists to run GSEA.
}
