% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{highlight_TF}
\alias{highlight_TF}
\title{Highlight certain transcription factors in a plotly graph.}
\usage{
highlight_TF(table, column, specialTF, markerColors)
}
\arguments{
\item{table}{Enrichment matrix/data.frame.}

\item{column}{Column # that stores the TF name in the matrix/df.}

\item{specialTF}{Named vector containing TF names as they appear in the
enrichment matrix/df and nicknames for their color group.
Example:
          specialTF<-c('HIF1A','EPAS1','ARNT','SIN3A')
          names(specialTF)<-c('HIF','HIF','HIF','SIN3A')}

\item{markerColors}{Vector specifying the shade for every color group.}
}
\value{
List of two objects:
A vector to attach to the enrichment matrix/df pointing out the color
group of every row.
A named vector connecting each color group to the chosen color.
}
\description{
Function to highlight certain transcription factors using
different colors in a plotly graph.
}
