% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{hypoxia_DESeq}
\alias{hypoxia_DESeq}
\title{RNA-Seq experiment}
\format{a DESeqResults objtec}
\usage{
hypoxia_DESeq
}
\description{
A DESeqResults objetc containing information of of an RNA-Seq
experiment on newly transcripted RNA in HUVEC cells during
two conditions, 8h of normoxia and 8h of hypoxia (deposited
at GEO as GSE89831).
}
\keyword{datasets}
