% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{log2.FC}
\alias{log2.FC}
\title{List of Entrez Gene IDs}
\format{
Array of 2754 log2(Fold Change) values.
}
\usage{
data("log2.FC")
}
\description{
Used to run examples. Array of 2754 log2(Fold Change) values
extracted from an RNA-Seq experiment.
}
\keyword{datasets}
