% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{makeTFBSmatrix}
\alias{makeTFBSmatrix}
\title{Function to search for a list of entrez gene IDs.}
\usage{
makeTFBSmatrix(GeneList, id_db, geneSetAsInput = TRUE)
}
\arguments{
\item{GeneList}{Array of gene Entrez IDs}

\item{id_db}{TF - gene binding database.}

\item{geneSetAsInput}{TRUE for lists of GeneSet objects,
FALSE for lists of vectors.}
}
\value{
1/0 matrix. Each row represents a gene, each column,
a ChIP-Seq file.
}
\description{
Function to search for a list of entrez gene IDs
in a  TF-gene binding data base.
}
\examples{
data('tfbs.database','Entrez.gene.IDs',package = 'TFEA.ChIP')
makeTFBSmatrix(Entrez.gene.IDs,tfbs.database)
}
