% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{metaanalysis_fx}
\alias{metaanalysis_fx}
\title{Perform Meta-analysis for each TF}
\usage{
metaanalysis_fx(dat)
}
\arguments{
\item{dat}{A data frame with columns: TF, OR, OR.SE, Accession, adj.pval.}
}
\value{
A list with:
  - summary: a data frame of ranked meta-analysis results per TF
  - results: a named list of raw meta-analysis objects from the 
   `meta` package
}
\description{
Conducts a random-effects meta-analysis of odds ratios (OR) 
and standard errors (OR.SE) for each TF using the `meta` package.
}
