% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{MetaData}
\alias{MetaData}
\title{TF-gene binding DB metadata}
\format{
A data frame of 1060 observations of 6 variables
}
\usage{
data("MetaData")
}
\description{
A data frame containing information about the ChIP-Seq experiments
used to build the TF-gene binding DB.
Fields in the data frame:
\itemize{
  \item Accession: Accession ID of the experiment.
  \item Cell: Cell line or tissue.
  \item 'Cell Type': More information about the cells.
  \item Treatment
  \item Antibody
  \item TF: Transcription factor tested in the ChIP-Seq experiment.
}
}
\keyword{datasets}
