% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{plot_ES}
\alias{plot_ES}
\title{Plots Enrichment Score from the output of GSEA.run.}
\usage{
plot_ES(
  GSEA_result,
  LFC,
  plot_title = NULL,
  specialTF = NULL,
  TF_colors = NULL,
  Accession = NULL,
  TF = NULL
)
}
\arguments{
\item{GSEA_result}{Returned by GSEA_run}

\item{LFC}{Vector with log2(Fold Change) of every gene that has an
Entrez ID. Arranged from higher to lower.}

\item{plot_title}{(Optional) Title for the plot}

\item{specialTF}{(Optional) Named vector of TF symbols -as written in
the enrichment table- to be highlighted in the plot. The name of each
element specifies its color group, i.e.: naming elements HIF1A and HIF1B
as 'HIF' to represent them with the same color.}

\item{TF_colors}{(Optional) Colors to highlight TFs chosen in specialTF.}

\item{Accession}{(Optional) restricts plot to the indicated list dataset
IDs.}

\item{TF}{(Optional) restricts plot to the indicated list transcription
factor names.}
}
\value{
Plotly object with a scatter plot -Enrichment scores- and a
heatmap -log2(fold change) bar-.
}
\description{
Function to plot the Enrichment Score of every member of
the ChIPseq binding database.
}
\examples{
data('GSEA.result','log2.FC',package = 'TFEA.ChIP')
TF.hightlight<-c('GATA2')
names(TF.hightlight)<-c('GATA2')
col<- c('red')
plot_ES(GSEA.result,log2.FC,specialTF = TF.hightlight,TF_colors = col)
}
