% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{plot_RES}
\alias{plot_RES}
\title{Plots all the RES stored in a GSEA_run output.}
\usage{
plot_RES(
  GSEA_result,
  LFC,
  plot_title = NULL,
  line.colors = NULL,
  line.styles = NULL,
  Accession = NULL,
  TF = NULL
)
}
\arguments{
\item{GSEA_result}{Returned by GSEA_run}

\item{LFC}{Vector with log2(Fold Change) of every gene that has an
Entrez ID. Arranged from higher to lower.}

\item{plot_title}{(Optional) Title for the plot.}

\item{line.colors}{(Optional) Vector of colors for each line.}

\item{line.styles}{(Optional) Vector of line styles for each line
('solid'/'dash'/'longdash').}

\item{Accession}{(Optional) restricts plot to the indicated list dataset
IDs.}

\item{TF}{(Optional) restricts plot to the indicated list transcription
factor names.}
}
\value{
Plotly object with a line plot -running sums- and a
heatmap -log2(fold change) bar-.
}
\description{
Function to plot all the RES stored in a GSEA_run output.
}
\examples{
data('GSEA.result','log2.FC',package = 'TFEA.ChIP')
plot_RES(GSEA.result,log2.FC,TF=c('GATA3'),
    Accession=c('ENCSR000BMX.GATA3.T47D',
    'ENCSR000EVW.GATA2.ENDOTHELIAL_UMBILICAL_VEIN'))
}
