% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{preprocessInputData}
\alias{preprocessInputData}
\title{Extracts data from a DESeqResults object or a data frame.}
\usage{
preprocessInputData(inputData, from.Mouse = FALSE)
}
\arguments{
\item{inputData}{DESeqResults object or data frame. In all cases
must include gene IDs. Data frame inputs should include 'pvalue' and
'log2FoldChange' as well.}

\item{from.Mouse}{TRUE if the input consists of mouse gene IDs.}
}
\value{
A table containing Entrez Gene IDs, LogFoldChange and p-val
values (both raw p-value and fdr adjusted p-value), sorted by
log2FoldChange.
}
\description{
Function to extract Gene IDs, logFoldChange, and p-val
values from a DESeqResults object or data frame. Gene IDs are
translated to ENTREZ IDs, if possible, and the resultant data frame
is sorted accordint to decreasing log2(Fold Change).
}
\examples{
data('hypoxia_DESeq',package='TFEA.ChIP')
preprocessInputData(hypoxia_DESeq)
}
