% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{SearchID}
\alias{SearchID}
\title{Searchs for an Entrez gene ID in a TF-gene database}
\usage{
SearchID(GeneID, id_db)
}
\arguments{
\item{GeneID}{gene Entrez ID.}

\item{id_db}{TF - gene binding database.}
}
\value{
1/0 row. Each element represents a TF ChIP-Seq experiment.
}
\description{
Searchs for an Entrez gene ID in a db that contains
the IDs of the genes a TF binds to.
}
