% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{tfbs.database}
\alias{tfbs.database}
\title{TFBS database for 3 ChIP-Seq datasets.}
\format{a data frame of 10 rows and 6 variables}
\usage{
data("tfbs.database")
}
\description{
Used to run examples. Output of the function GR2tfbs_db from the
TFEA.ChIP package. Contains a list of three vectors of Entrez Gene
IDs assoctiated to three ChIP-Seq experiments
}
\keyword{datasets}
