% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{ChIPDB}
\alias{ChIPDB}
\title{TF-Gene List}
\format{
A list with two elements:
Gene Keys, a vector of Entrez IDs
ChIP Targets, a list containing indices of Entrez IDs associated with peaks from various ChIP-Seq experiments.
}
\usage{
data("ChIPDB")
}
\description{
This dataset contains two elements: the first is "Gene Keys," which
includes all the Entrez IDs, and the second is "ChIP Targets," a list
containing information from multiple ChIP-Seq experiments. Each entry
in this list contains the indices of the Entrez IDs from the first
element that are associated with the peaks of that specific ChIP-Seq.
}
\keyword{datasets}
