% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{contingency_matrix}
\alias{contingency_matrix}
\title{Computes 2x2 contingency matrices}
\usage{
contingency_matrix(test_list, control_list,
  chip_index = get_chip_index())
}
\arguments{
\item{test_list}{List of gene Entrez IDs}

\item{control_list}{If not provided, all human genes not present in
test_list will be used as control.}

\item{chip_index}{Output of the function “get_chip_index”, a data frame
containing accession IDs of ChIPs on the database and the TF each one
tests. If not provided, the whole internal database will be used}
}
\value{
List of contingency matrices, one CM per element in chip_index
(i.e. per ChIP-seq dataset).
}
\description{
Function to compute contingency 2x2 matrix by the partition
of the two gene ID lists according to the presence or absence of the
terms in these list in a ChIP-Seq binding database.
}
\examples{
data('Genes.Upreg',package = 'TFEA.ChIP')
CM_list_UP <- contingency_matrix(Genes.Upreg)
}
