% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{filter_expressed_TFs}
\alias{filter_expressed_TFs}
\title{Filter Expressed TFs}
\usage{
filter_expressed_TFs(Table, chip_index, TFfilter = NULL, encodeFilter = FALSE)
}
\arguments{
\item{Table}{A data frame containing gene expression data with a `Genes` column.}

\item{chip_index}{A data frame containing ChIP-Seq dataset accession IDs 
and associated TFs.}

\item{TFfilter}{(Optional) A character vector of TFs to filter.}

\item{encodeFilter}{(Optional) Logical; if TRUE, applies ENCODE filtering 
to ChIP-Seq data.}
}
\value{
A filtered `chip_index` data frame containing only expressed TFs.
}
\description{
Filters TFs based on their expression status in the input dataset.
This function identifies expressed TFs by intersecting the input gene list
with the `chip_metadata` dataset.
}
