% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-functions.R
\name{highlight_TF}
\alias{highlight_TF}
\title{Highlight special TFs in enrichment table.}
\usage{
highlight_TF(enrichTab, column, specialTF)
}
\arguments{
\item{enrichTab}{Data frame containing the enrichment results.}

\item{column}{The column index or name of the enrichment table used 
for matching TFs.}

\item{specialTF}{A named vector of transcription factors to highlight
in the plot.}
}
\value{
A list containing the updated highlight column and the color 
mapping for each TF.
}
\description{
Function to assign special TFs to specific colors and 
highlight them in the enrichment table.
}
