% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{stat_mat}
\alias{stat_mat}
\title{Data frame, output from the function getCMstats from the TFEA.ChIP package}
\format{a data frame of 10 rows and 6 variables}
\usage{
data("stat_mat")
}
\description{
Used to run examples. Output of the function getCMstats from the TFEA.ChIP package,
is a data frame storing the following fields:
\itemize{
  \item Accession: GEO or Encode accession ID for each ChIP-Seq dataset.
  \item TF: Transcription Factor tested.
  \item p.value: p-value of the Fisher test performed on a contingency matrix for each ChIP-Seq experiment.
  \item OR: Odds Ratio on the contingency matrix done for each ChIP-Seq experiment.
  \item adj.p.value: p-value adjusted by FDR
  \item log.adj.pVal: log10(adj.p.value)
}
}
\keyword{datasets}
